FlashCard Generator/requirements.txt
# requirements.txt
streamlit==1.34.0
openai==1.30.1
pypdf==4.2.0
python-dotenv==1.0.1